+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Section Order:
 
 - `views.ajax_fleet_details` refactored
 - `tasks._process_fleet` refactored
+- `tasks.open_fleet` refactored
 
 ## [2.6.0] - 2025-08-14